+++ OpenDiscordHaxx/Config.cs
@@ -1,4 +1,6 @@
 ﻿using Newtonsoft.Json;
+using System;
+using System.IO;
 
 namespace DiscordHaxx
 {
@@ -7,8 +9,18 @@ namespace DiscordHaxx
         [JsonProperty("enable_gateway")]
         public bool EnableGateway { get; private set; }
 
-        
+
         [JsonProperty("gateway_cap")]
         public int GatewayCap { get; private set; }
+
+        public Config()
+        {
+            File.WriteAllText("Config.json", JsonConvert.SerializeObject(this));
+        }
+        
+        [JsonConstructor]
+        [Obsolete("This is not meant to be called, its just here to troll the deserializer so it doesnt create the config file twice bc i cba to fucking make an if statement", true)]
+        public Config(bool we_do_a_little_trolling)
+        {}
     }
 }