+++ OpenDiscordHaxx/Tools/Recon/Requests/Opcode.cs
@@ -5,6 +5,7 @@
         Id,
         StartRecon,
         ReconCompleted,
-        ReconFailed
+        ReconFailed,
+        NoId
     }
 }